refactor : Step 2 패키지 위치 변경
