+++ src/main/java/study/calculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
