+++ src/test/java/study/calculator/TestStringAddCalculator.java
@@ -1,11 +1,15 @@
-package study;
+package study.calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
 
+import study.calculator.StringAddCalculator;
+
+
 public class TestStringAddCalculator {
+    
     @Test
     public void splitAndSum_null_또는_빈문자() {
         int result = StringAddCalculator.splitAndSum(null);
